crm-futures-project/backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
Flask-Cors>=3.0
python-dotenv>=0.19
mysql-connector-python>=8.0
gunicorn>=20.1
Werkzeug>=2.0
Flask-JWT-Extended>=4.0 